# lib/api-zod/src/generated/types/adInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sarad streaming platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdInputType } from "./adInputType";

export interface AdInput {
  title: string;
  titleAr?: string;
  type: AdInputType;
  imageUrl?: string;
  videoUrl?: string;
  linkUrl?: string;
  isActive?: boolean;
  order?: number;
}