# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sarad streaming platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginInput {
  email: string;
  password: string;
}